Mesecons work for all door doors.
Mesecons were set to open and close wooden doors, and steel doors. But there are now glass doors in the doors mod, and there are also the doors in homedecor which may one day use the doors mod to handle their door-ish-ness. So instead of continuing to hardcode more stuff from other modules, simply add mesecons support to all doors.

Also since mesecons_compatibility is only dealing with doors in the first place, just rename it to mesecons_doors.