Allowed loops & fixed certain strings
I've added a simple lexer to work out the correct places to add dummy
code, which fixes the problem mentioned in code_prohibited.
This means LuaControllers are no longer restricted to simple
conditionals, but should still be safe for server owners.
Furthermore, the lexer prevents accidentally matching text in strings.