Prevent attack due to strings being completely removed.
This was a viable attack:
function x() end x"x"repeat x"x"until false

The reason is that removing the quotes caused the x and the keyword to be joined, causing them not to be found. Fixed by replacing the string with a space. It's not necessary for comments because the -- already stays.