[mesecons_doors] Improve code quality and fix Voxelgarden support
The main code quality improvement here is the functionalization of code:
now trapdoors are also overriden using their own function, and both
meseconify_door and meseconify_trapdoor are now ran inside for loops
with local lists of base node names. In the future, this could help
refactor this mod entirely to not list doors manually but parse through
the doors.registered_* variables, ensuring that every door works.

As a nice bonus, Voxelgarden support is fixed, and potentially so is
support for other games with exotic implementations of the doors mod. No
longer are we assuming that the existence of doors.get means that all
the trapdoors exist too; now, if the meseconify function doesn't find
their definition, it just returns without crashing.